hydrology: make moisture/precip default and rename legacy
