+++ mods/mod-swooper-maps/src/domain/hydrology/ops/compute-precipitation/strategies/index.ts
@@ -1,3 +1,3 @@
-export { defaultStrategy } from "./baseline.js";
-export { vectorStrategy } from "./vector.js";
+export { basicStrategy } from "./baseline.js";
+export { defaultStrategy } from "./vector.js";
 export { refineStrategy } from "./refine.js";
